Rename `classThread.py` -> `thread.py`.
